pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-landscrap"
version = "0.1.0"
description = "Turn deleted code into generative software art artifacts."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "code-landscrap" }]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "google-genai>=0.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
code-landscrap = "code_landscrap.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
code_landscrap = ["templates/*.html", "templates/*.css", "templates/*.js"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -ra"
